Update PersonalData.cs
fixed

+++ src/UMS.Domain/Entities/User/PersonalData.cs
@@ -7,8 +7,8 @@ namespace UMS.Domain.Entities.People
         public string FirstName { get; set; }
         public string MiddleName { get; set; }
         public string LastName { get; set; }
-        public int CityId { get; set; }
-        public int CountryId { get; set; }
+        public long CityId { get; set; }
+        public long CountryId { get; set; }
         public string Email { get; set; }
         public Gender Gender { get; set; }
         public string PhoneNumber { get; set; }
